Move webapp config to separate file, add entity scan
